+++ config/config.php
@@ -1,21 +1,53 @@
 <?php
 
 return [
-    'eddbk' => [
-        'week_starts_on'        => 'sunday',
-        'default_calendar_view' => 'week',
-        'booking_wizard_color'  => '#17a7dd',
-        'booking_statuses_colors'     => [
-            'draft'     => '#dfe4ea',
-            'pending'   => '#1e90ff',
-            'scheduled' => '#2ed573',
-            'completed' => '#57606f',
-            'cancelled' => '#eb4d4b',
-        ],
-    ],
     'wp_bookings_ui' => [
         'wp_rest_api_nonce' => 'wp_rest',
         'settings' => [
+            'wizard_labels' => [
+                'general' => [
+                    'title' => 'Book an Appointment',
+                    'buttons' => [
+                        'back' => 'Back',
+                        'next' => 'Next',
+                        'book' => 'Book Now',
+                    ],
+                ],
+                'preview' => [
+                    'price' => 'Starting at %s for a %s appointment.',
+                    'available' => 'More appointment durations available in the next step.',
+                    'booking' => 'You are booking:',
+                ],
+                'confirmation' => [
+                    'booking' => 'You are booking: *%s* - %s appointment',
+                    'starting' => 'Starting at *%s*',
+                    'price' => 'The price is of *%s*',
+                ],
+                'fields' => [
+                    'service' => [
+                        'title' => 'Select a service',
+                        'placeholder' => 'Select option',
+                    ],
+                    'duration' => [
+                        'title' => 'Select duration',
+                    ],
+                    'date' => [
+                        'title' => 'Select a date',
+                        'empty' => 'No appointments are available this month.',
+                    ],
+                    'time' => [
+                        'title' => 'Select available time',
+                        'placeholder' => 'Select a date to pick a time from.',
+                    ],
+                    'timezone' => [
+                        'title' => 'Change timezone',
+                    ],
+                    'notes' => [
+                        'title' => 'Additional notes',
+                        'placeholder' => 'If you have got any special requests for this service, please note them down here.',
+                    ],
+                ],
+            ],
             'options' => [
                 'week_starts_on' => [
                     'sunday' => 'Sunday',
@@ -30,15 +62,29 @@ return [
                     'day'   => 'Day',
                     'week'  => 'Week',
                     'month' => 'Month',
-                ]
+                ],
             ],
             'fields' => [
                 'week_starts_on',
                 'default_calendar_view',
                 'booking_wizard_color',
                 'booking_statuses_colors',
+                'booking_wizard_labels'
             ],
             'prefix' => 'eddbk',
+            'default_values' => [
+                'week_starts_on'        => 'sunday',
+                'default_calendar_view' => 'week',
+                'booking_wizard_color'  => '#17a7dd',
+                'booking_statuses_colors'     => [
+                    'draft'     => '#dfe4ea',
+                    'pending'   => '#1e90ff',
+                    'scheduled' => '#2ed573',
+                    'completed' => '#57606f',
+                    'cancelled' => '#eb4d4b',
+                ],
+                'booking_wizard_labels' => []
+            ],
             'array_fields' => [
                 'booking_statuses_colors'
             ],