+++ src/Handlers/GeneralUiStateHandler.php
@@ -15,10 +15,10 @@ use Dhii\Util\Normalization\NormalizeArrayCapableTrait;
 use Dhii\Util\Normalization\NormalizeIterableCapableTrait;
 use Dhii\Util\Normalization\NormalizeStringCapableTrait;
 use Dhii\Util\String\StringableInterface as Stringable;
-use InvalidArgumentException;
 use Psr\Container\ContainerInterface;
 use Psr\EventManager\EventInterface;
 use Psr\EventManager\EventManagerInterface;
+use RebelCode\Bookings\WordPress\Module\IteratorToArrayRecursiveCapableTrait;
 use RebelCode\Modular\Events\EventsConsumerTrait;
 use stdClass;
 use Traversable;
@@ -63,6 +63,9 @@ class GeneralUiStateHandler implements InvocableInterface
 
     /* @since [*next-version*] */
     use NormalizeIterableCapableTrait;
+
+    /* @since [*next-version*] */
+    use IteratorToArrayRecursiveCapableTrait;
 
     /**
      * Settings container.
@@ -355,31 +358,5 @@ class GeneralUiStateHandler implements InvocableInterface
         }
 
         return $preparedLinksConfig;
-    }
-
-    /**
-     * Convert an iterator to an array.
-     *
-     * @since [*next-version*]
-     *
-     * @param array|stdClass|Traversable $iterator The object to convert.
-     *
-     * @throws InvalidArgumentException If $iterator is not iterable.
-     *
-     * @return array Result of iterator to array transform.
-     */
-    protected function _iteratorToArrayRecursive($iterator)
-    {
-        $iterator = $this->_normalizeIterable($iterator);
-        $array    = [];
-        foreach ($iterator as $key => $value) {
-            if ($value instanceof Traversable || is_array($value) || $value instanceof stdClass) {
-                $array[$key] = $this->_iteratorToArrayRecursive($value);
-                continue;
-            }
-            $array[$key] = $value;
-        }
-
-        return $array;
     }
 }