+++ src/WpBookingsUiModule.php
@@ -142,6 +142,8 @@ class WpBookingsUiModule extends AbstractBaseModule
 
         $this->_attach('eddbk_settings_ui_state', $c->get('eddbk_settings_ui_state_handler'));
 
+        $this->_attach('eddbk_front_application_labels', $c->get('eddbk_front_application_labels_handler'));
+
         $this->_attach('wp_ajax_set_' . $c->get('wp_bookings_ui/screen_options/key'), $c->get('eddbk_bookings_save_screen_options_handler'));
 
         $this->_attach('wp_ajax_' . $c->get('wp_bookings_ui/settings/action'), $c->get('eddbk_bookings_update_settings_handler'));
@@ -478,9 +480,11 @@ class WpBookingsUiModule extends AbstractBaseModule
                 $settingsTemplate = $c->get('eddbk_ui_settings_template');
 
                 $generalSettingsTabContent = $c->get('eddbk_ui_settings_general_tab_template')->render();
+                $wizardSettingsTabContent = $c->get('eddbk_ui_settings_wizard_tab_template')->render();
                 $componentsContent = $this->_renderTemplate('components');
 
                 echo $settingsTemplate->render([
+                    'wizardSettingsTab' => $wizardSettingsTabContent,
                     'generalSettingsTab' => $generalSettingsTabContent,
                     'components'         => $componentsContent,
                 ]);